refactor dubbo-samples-metrics-prometheus
