+++ 4-governance/dubbo-samples-metrics-prometheus/dubbo-samples-metrics-prometheus-consumer/src/main/java/org/apache/dubbo/samples/metrics/prometheus/consumer/MetricsConsumer.java
@@ -1,0 +1,51 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.dubbo.samples.metrics.prometheus.consumer;
+
+import org.apache.dubbo.config.annotation.DubboReference;
+import org.apache.dubbo.config.spring.context.annotation.EnableDubbo;
+import org.apache.dubbo.samples.metrics.prometheus.api.DemoService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.ConfigurableApplicationContext;
+
+@EnableDubbo
+@SpringBootApplication
+public class MetricsConsumer {
+
+    private static Logger logger = LoggerFactory.getLogger(MetricsConsumer.class);
+
+    @DubboReference
+    private DemoService demoService;
+
+    public static void main(String[] args) {
+        ConfigurableApplicationContext ctx = SpringApplication.run(MetricsConsumer.class, args);
+        DemoService demoService = ctx.getBean(DemoService.class);
+        while (true) {
+            try {
+                Thread.sleep(3000);
+                System.out.println(demoService.sayHello("Dubbo").getMsg());
+            } catch (InterruptedException e) {
+                logger.error("MetricsConsumer failed: ", e);
+            }
+        }
+    }
+
+}
